Convert song ID from cuid to int
The song id will be changed from

  id          String  @id @default(cuid())

  to

    id          Int     @id @default(autoincrement())

Mainly what we're after here is having shorter IDs. Could have done some
sophisticated ID generation at 4 characters but at the end of the day
it doesn't really matter if people know the order. Part of the project was
to expose an API so having these exposed is trivial