Acquire songs and filter client side
Previously we were using prisma filtering. The main problem was with
the inconsistent state twitch in the input as it would adjust. Typing cow
left the user with "cw". This seemed to be particulare to the large dataset
and SSR but I'm not sure.

At the end of the day it was actually better anyhow to do all filtering
client side as we will likely not have more than 300 records for a while.
The current count is <250.

Also in this PR is the introduction of fuzzy search added by fuse js
I lowered the threshold a little bit but the main target was to match
"kesha" to "ke$ha". 0.6 seemed way too loose so I dropped to 0.3.
Somewhere between 0.3 and 0.4 was good so I went with 0.35